alpha-pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alpha-pipeline"
version = "1.0.0"
description = "Enterprise-grade equity alpha generation pipeline"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Alpha Pipeline Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3.11",
    "Operating System :: MacOS :: MacOS X",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "pandas>=2.2.2",
    "numpy>=1.26.4",
    "scipy>=1.13.1",
    "scikit-learn>=1.5.1",
    "lightgbm>=4.4.0",
    "pytorch-forecasting>=1.1.0",
    "pytorch-lightning>=2.3.3",
    "darts>=0.29.0",
    "yfinance>=0.2.40",
    "pandas-ta>=0.3.14b",
    "pyarrow>=16.1.0",
    "pydantic>=2.8.2",
    "pyyaml>=6.0.1",
    "matplotlib>=3.9.1",
    "seaborn>=0.13.2",
    "plotly>=5.22.0",
    "tqdm>=4.66.4",
    "structlog>=24.4.0",
    "numba>=0.60.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.2.2",
    "requests>=2.32.3",
    "joblib>=1.4.2",
    "bottleneck>=1.4.0",
    "cachetools>=5.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.2",
    "pytest-cov>=5.0.0",
    "black>=24.4.2",
    "isort>=5.13.2",
    "mypy>=1.10.1",
    "ruff>=0.5.0",
]
dashboard = [
    "streamlit>=1.36.0",
    "dash>=2.17.1",
]
alerts = [
    "slack-sdk>=3.31.0",
]

[project.scripts]
alpha-pipeline = "src.run_pipeline:main"
alpha-live = "src.inference.live_loader:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "N", "B", "C4"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
